refactor: Компоненты разделены на отдельные файлы
