+++ src/components/burger-Ingredients/burger-ingredients.module.css
@@ -9,39 +9,6 @@
     max-height: 80%;
 }
 
-.container {
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    width: 272px;
-    height: 208px;
-    align-items: center;
-    justify-content: center;
-}
-
-.image {
-    object-fit: cover;
-}
-
-.textBox {
-    box-sizing: border-box;
-    padding-bottom: 20px;
-    text-align: center;
-}
-
-.price {
-    display: flex;
-    gap: 10px;
-    justify-content: center;
-    padding: 8px;
-}
-
-.typeBox {
-    display: flex;
-    align-content: start;
-    flex-wrap: wrap;
-}
-
 .navBar {
     display: flex;
 }
