+++ src/components/burger-Ingredients/items-block/items-block.module.css
@@ -1,0 +1,5 @@
+.typeBox {
+    display: flex;
+    align-content: start;
+    flex-wrap: wrap;
+}
